backend/requirements.txt
Django==3.2.23
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
djongo==1.3.6
pymongo==3.12.3
python-decouple==3.8
python-dotenv==1.1.0
firebase-admin==6.2.0
django-ratelimit==4.1.0
gunicorn==21.2.0
whitenoise==6.6.0
sqlparse==0.2.4
